+++ dialect/mysql/select_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/antlr/antlr4/runtime/Go/antlr/v4"
 	"github.com/stephenafamo/bob/dialect/mysql"
+	"github.com/stephenafamo/bob/dialect/mysql/fm"
 	"github.com/stephenafamo/bob/dialect/mysql/sm"
 	testutils "github.com/stephenafamo/bob/test/utils"
 	mysqlparser "github.com/stephenafamo/sqlparser/mysql"
@@ -49,12 +50,9 @@ func TestSelect(t *testing.T) {
 				sm.From(mysql.Select(
 					sm.Columns(
 						"status",
-						mysql.F("LEAD", "created_date", 1, mysql.F("NOW")).
-							Over().
-							PartitionBy("presale_id").
-							OrderBy("created_date").
-							Minus(mysql.Quote("created_date")).
-							As("difference")),
+						mysql.F("LEAD", "created_date", 1, mysql.F("NOW"))(
+							fm.Over().PartitionBy("presale_id").OrderBy("created_date"),
+						).Minus(mysql.Quote("created_date")).As("difference")),
 					sm.From("presales_presalestatus")),
 				).As("differnce_by_status"),
 				sm.Where(mysql.Quote("status").In(mysql.S("A"), mysql.S("B"), mysql.S("C"))),