+++ dialect/sqlite/dialect/function.go
@@ -8,8 +8,11 @@ import (
 	"github.com/stephenafamo/bob/expr"
 )
 
-func NewFunction(name string, args ...any) Function {
-	return Function{name: name, args: args}
+func NewFunction(name string, args ...any) *Function {
+	f := &Function{name: name, args: args}
+	f.Chain = expr.Chain[Expression, Expression]{Base: f}
+
+	return f
 }
 
 type Function struct {
@@ -17,30 +20,17 @@ type Function struct {
 	args []any
 
 	// Used in value functions. Supported by Sqlite and Postgres
-	filter []any
+	Distinct bool
+	clause.OrderBy
+	Filter []any
+	w      *clause.Window
 
 	// For chain methods
 	expr.Chain[Expression, Expression]
 }
 
-// A function can be a target for a query
-func (f *Function) Apply(q *clause.From) {
-	q.Table = f
-}
-
-func (f *Function) Filter(e ...any) *Function {
-	f.filter = append(f.filter, e...)
-
-	return f
-}
-
-func (f *Function) Over() *functionOver {
-	fo := &functionOver{
-		function: f,
-	}
-	fo.WindowChain = &WindowChain[*functionOver]{Wrap: fo}
-	fo.Base = fo
-	return fo
+func (f *Function) SetWindow(w clause.Window) {
+	f.w = &w
 }
 
 func (f Function) WriteSQL(w io.Writer, d bob.Dialect, start int) ([]any, error) {
@@ -50,37 +40,36 @@ func (f Function) WriteSQL(w io.Writer, d bob.Dialect, start int) ([]any, error)
 
 	w.Write([]byte(f.name))
 	w.Write([]byte("("))
+
+	if f.Distinct {
+		w.Write([]byte("DISTINCT "))
+	}
+
 	args, err := bob.ExpressSlice(w, d, start, f.args, "", ", ", "")
 	if err != nil {
 		return nil, err
 	}
+
+	orderArgs, err := bob.ExpressIf(w, d, start+len(args), f.OrderBy,
+		len(f.OrderBy.Expressions) > 0, " ", "")
+	if err != nil {
+		return nil, err
+	}
+	args = append(args, orderArgs...)
+
 	w.Write([]byte(")"))
 
-	filterArgs, err := bob.ExpressSlice(w, d, start, f.filter, " FILTER (WHERE ", " AND ", ")")
+	filterArgs, err := bob.ExpressSlice(w, d, start, f.Filter, " FILTER (WHERE ", " AND ", ")")
 	if err != nil {
 		return nil, err
 	}
 	args = append(args, filterArgs...)
 
+	winargs, err := bob.ExpressIf(w, d, start+len(args), f.w, f.w != nil, "OVER (", ")")
+	if err != nil {
+		return nil, err
+	}
+	args = append(args, winargs...)
+
 	return args, nil
-}
-
-type functionOver struct {
-	function *Function
-	*WindowChain[*functionOver]
-	expr.Chain[Expression, Expression]
-}
-
-func (wr *functionOver) WriteSQL(w io.Writer, d bob.Dialect, start int) ([]any, error) {
-	fargs, err := bob.Express(w, d, start, wr.function)
-	if err != nil {
-		return nil, err
-	}
-
-	winargs, err := bob.ExpressIf(w, d, start+len(fargs), wr.def, true, "OVER (", ")")
-	if err != nil {
-		return nil, err
-	}
-
-	return append(fargs, winargs...), nil
 }