+++ dialect/sqlite/sm/qm.go
@@ -63,12 +63,12 @@ func GroupBy(e any) bob.Mod[*dialect.SelectQuery] {
 	}
 }
 
-func Window(name string) dialect.WindowMod[*dialect.SelectQuery] {
-	m := dialect.WindowMod[*dialect.SelectQuery]{
+func Window(name string) dialect.WindowsMod[*dialect.SelectQuery] {
+	m := dialect.WindowsMod[*dialect.SelectQuery]{
 		Name: name,
 	}
 
-	m.WindowChain = &dialect.WindowChain[*dialect.WindowMod[*dialect.SelectQuery]]{
+	m.WindowChain = &dialect.WindowChain[*dialect.WindowsMod[*dialect.SelectQuery]]{
 		Wrap: &m,
 	}
 	return m